Access to resources.

+++ RestaurantApiUdemyCS6/Services/RestaurantService.cs
@@ -1,18 +1,21 @@
 ﻿using AutoMapper;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.EntityFrameworkCore;
+using RestaurantApiUdemyCS6.Authorization;
 using RestaurantApiUdemyCS6.Entities;
 using RestaurantApiUdemyCS6.Exceptions;
 using RestaurantApiUdemyCS6.Models;
+using System.Security.Claims;
 
 namespace RestaurantApiUdemyCS6.Services
 {
-    public interface IRestaurantService 
+    public interface IRestaurantService
     {
         RestaurantDto GetById(int id);
         IEnumerable<RestaurantDto> GetAll();
-        int Create(CreateRestaurantDto dto);
-        void Delete(int id);
-        void Update(int id, UpdateRestaurantDto dto);
+        int Create(CreateRestaurantDto dto, int userId);
+        void Delete(int id, ClaimsPrincipal user);
+        void Update(int id, UpdateRestaurantDto dto, ClaimsPrincipal user);
     }
 
     public class RestaurantService : IRestaurantService
@@ -20,41 +23,61 @@ namespace RestaurantApiUdemyCS6.Services
         private readonly RestaurantDbContext _dbContext;
         private readonly IMapper _mapper;
         private readonly ILogger _logger;
+        private readonly IAuthorizationService _authorizationService;
 
-        public RestaurantService(RestaurantDbContext dbContext, IMapper mapper, ILogger<RestaurantService> logger)
+        public RestaurantService(RestaurantDbContext dbContext,
+            IMapper mapper, ILogger<RestaurantService> logger,
+            IAuthorizationService authorizationService)
         {
             _dbContext = dbContext;
             _mapper = mapper;
             _logger = logger;
+            _authorizationService = authorizationService;
         }
 
-        public void Update(int id, UpdateRestaurantDto dto)
+        public void Update(int id, UpdateRestaurantDto dto, ClaimsPrincipal user)
         {
             var restaurantToUpdate = _dbContext
-                .Restaurants
-                .FirstOrDefault(r => r.Id == id);
+            .Restaurants
+            .FirstOrDefault(r => r.Id == id);
 
             if (restaurantToUpdate is null)
                 throw new NotFoundException("Restaurant not found");
 
+            var authorizationResult = _authorizationService.AuthorizeAsync(user,
+                restaurantToUpdate, new ResourceOperationRequirement(ResourceOperation.Update)).Result;
+
+            if (!authorizationResult.Succeeded)
+            {
+                throw new ForbidException();
+            }
+
             restaurantToUpdate.Name = dto.Name;
-            restaurantToUpdate.Description= dto.Description;
+            restaurantToUpdate.Description = dto.Description;
             restaurantToUpdate.HasDelivery = dto.HasDelivery;
 
             _dbContext.SaveChanges();
         }
 
-        public void Delete(int id)
+        public void Delete(int id, ClaimsPrincipal user)
         {
             _logger.LogError($"Restaurant with id: {id} DELETE action invoked");
 
             var restaurant = _dbContext
                   .Restaurants
                   .FirstOrDefault(r => r.Id == id);
-            
-            if(restaurant is null)
+
+            if (restaurant is null)
             {
                 throw new NotFoundException("Restaurant not found");
+            }
+
+            var authorizationResult = _authorizationService.AuthorizeAsync(user,
+            restaurant, new ResourceOperationRequirement(ResourceOperation.Delete)).Result;
+
+            if (!authorizationResult.Succeeded)
+            {
+                throw new ForbidException();
             }
 
             _dbContext.Restaurants.Remove(restaurant);
@@ -91,9 +114,10 @@ namespace RestaurantApiUdemyCS6.Services
             return restaurantDtos;
         }
 
-        public int Create(CreateRestaurantDto dto)
+        public int Create(CreateRestaurantDto dto, int userId)
         {
             var restaurant = _mapper.Map<Restaurant>(dto);
+            restaurant.CreatedById = userId;
             _dbContext.Restaurants.Add(restaurant);
             _dbContext.SaveChanges();
